refactor(rename): перенести RenameWorkspaceEditBuilder в пакет rename и кэшировать флаги внутри
Перенос RenameWorkspaceEditBuilder из пакета providers в новый подпакет
rename (класс + ресурсные бандлы _ru/_en, добавлен @NullMarked package-info).
Кэширование флагов documentChangesSupport/changeAnnotationSupport перенесено
внутрь билдера: инжектирован ClientCapabilitiesHolder, добавлены поля и
@EventListener handleInitializeEvent по образцу DefinitionProvider; build()
больше не принимает флаги параметрами. RenameProvider избавлен от этих полей
и собственного обработчика initialize.
