+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/rename/RenameWorkspaceEditBuilder_en.properties
